pyproject.toml
[tool.poetry]
name = "django-clean-code"
version = "0.1.0"
description = ""
authors = ["Azhar Iqbal <azhariqbal2942@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
pyclean = "^2.7.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
